Add timestamp header for New Relic request queue (#1672)
As discussed in https://github.com/linkerd/linkerd/pull/1664#issuecomment-335031008, the New Relic request queue uses a `X-Request-Start` header with a timestamp to measure latency. I've added a filter that allows Linkerd to add such a header to outbound requests. I've added unit and end-to-end tests for this feature, and added the header configuration to the config docs.