Stop k8s watches from time travelling (#1681)
I've modified our code for watching Kubernetes resources so that it will ignore events whose resource versions are older than the resource version of the previous event. We do this by creating a two-item sliding window over the stream of watch events, and filtering out those tuples where the first item (the previous event) is older than the second item (the received event). This should ensure that we never travel back in time when the Kubernetes API is out of sync.

I've added a test for `EndpointsNamer` and `ServiceNamer` that assert that the namer's state does not update when receiving events in reverse order. Furthermore, a number of existing tests had to be modified to ensure all resource versions were sequential – we probably had been lazy in the past about copying and pasting JSON strings in tests without updating their resource versions.

Closes #1680