Denoise Consul namer logging (#1682)
As described in issue #1670, the Consul namer currently logs a scary-looking error message every time we close a watch on a Consul service. This is because the namer's activity handles the `ServiceRelease` exception that's thrown to close a watch that's no longer needed as though it were an error. 

I've modified `SvcAddr` to handle the `ServiceRelease` exception separately from other `Throw`s, which should stop the spurious error message from being logged.

Fixes #1670