Re-initialize namerd clients on SIGHUP
Users have reported an issue where linkerd instances may stop receiving
updates from namerd.

While we debug this more fully, it's useful to have an escape-hatch to
simply force linkerd to reload its namerd clients.

This change introduces a global singleton Event/Witness,
InterpreterRefresher, which is notified when SIGHUP is handled. This
causes a new namerd client to be provided to the thrift namerd
interpreter, and all activities that had been built on the old client
are rebuilt on the new client.

This has been tested locally.