Add test for TLS certificate reloading (#1692)
I've added a test class to `linkerd/protocol/http/integration` which exercises the reloading of TLS certificates. The test prepares 2 different cert-key pairs: one with `commonName = foo `and one with `commonName = bar`. It configures an incoming Linkerd router with server TLS and points it at the "foo" cert and key, and an outgoing Linkerd router with client TLS and set it to expect `commonName = bar`.

The test sends a request through and asserts that it should fail validation because the cert has commonName foo instead of bar. Then, it replaces the bar cert and key with the foo cert and key, and sends another request, asserting that this request succeeds. If the incoming Linkerd reloads the cert and key, the second request should succeed.