Properly handle cancellation in ThirftNamerInterface (#1698)
When a client establishes a pending request on Namerd's ThirftNamerInterface and then disconnects, the request stays pending rather than being properly cancelled.  If this occurs repeatedly, it can lead to a memory leak in Namerd as the number of pending requests grows without bound.

Add an interrupt handler to the Promise returned by Observer so that the Future fails when it is cancelled.

Manually tested that the pending requests stat properly resets when the client disconnects.