Added a namer for Rancher
Linkerd currently does not support service-discovery in environments that uses Rancher for orchestration. This makes Linkerd less attractive to people using Rancher for the Docker environments, as they would have to implement additional infrastructure to get service-discovery that way. Rancher has a JSON-based metadata-API that contains information about all of the containers that are running in Rancher, so it would be possible to perform service-discovery that way.

This commit introduces a new namer that utilizes the long-polling functionality of the Rancher metadata-API to achieve near-instant updates to the discovered services.

I have run some small-scale validation tests on a Rancher platform to ensure that the client accuratly detects the running services and can route requests to them.

Fixes #1146

Signed-off-by: Morten Fangel <fangel@sevengoslings.net>