Fix incorrect handling of Consul ServiceRelease exception (#1738)
In the Consul namer `SvcAddr` class, a singleton exception called `ServiceRelease` is raised when an `Activity` watching Consul is closed. Previously, this exception was a `val`. However, due to how equality is implemented for `Failure`s, the by the time exception raised reached the `match` in `loop()`, the `Failure` did not always equal the value of `ServiceRelease`, seemingly because of the `ProcessFailure` filter masking some of the flags on the `Failure`. This led to service releases being handled as errors, rather than gracefully cancelling the observation, leading to a _lot_ of unnecessary exceptions in the logs.

I've fixed this issue by setting the cause of the raised exception to a singleton object rather than a `val`, so that the match arm can test type rather than equality. The exception raised to cancel an observation should now always match the correct pattern.

I tried to write a unit test for this that makes assertions about the contents of the logs, but I couldn't make it fail against the current master. It appeared that the failure was being munged only when the namer was fully integrated with Linkerd. Instead, I tested this manually using Consul running locally and setting the number of bound names in the binding cache to 1. I would request one service in Consul, and then a second one – thus forcing the namer to evict the first bound name from the bindcache, stopping the observation.

Fixes #1670, for real this time.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>