Refactor ApiTest to use AsyncQueue and be less flakey (#1747)
The use of ConcurrentLinkedQueue was race-y because one thread would be populating the queue (the watchable) while the main test thread polled and tested the queue. If the poll occurs before the queue is populated, the test will fail.

We replace CLQ with AsyncQueue and use await on each poll(). That way, if the poll occurs before the queue is populated, the await will block until the value is available.

Signed-off-by: Alex Leong <alex@buoyant.io>