Add a check to disallow invalid namespaces (#1739)
Problem
When namerd uses consul for dtab storage, invalid characters in a namespace e.g. %2e%2e/foo causes namerd to infinitely log a dtab observation error

Solution
Add a check to ensure that only alphanumeric characters can be used as keys in consul to store dtabs.

Validation
Created a varying number of valid and invalid strings with a regex pattern and curled namerd with both types of strings. The tests worked successfully. However, curling a large number of requests to namerd caused consul to remove namerd as a service and subsequently caused ChannelClosedExceptions. I believe that is probably caused by DoSing consul with a bash script and not because of the proposed change.

Fixes #1733

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>