Improve handling of dtab parse errors in ConsulDtabStore (#1760)
Fixes #1759.

Currently, when Namerd's ConsulDtabStore makes a request to Consul's KV API that returns a value, it passes that value to Dtab.read() to parse it as a dtab. Dtab parsing can fail if the string returned by the KV API is not properly formatted as a dtab; if this is the case, Dtab.read() will throw an InvalidArgumentException. However, this is not handled in ConsulDtabStore; instead, the exception will be thrown upwards. When the dtab observation was opened by the Admin UI's DtabHandler, no additional error handling is performed, causing the exception to cancel the request, printing a message like

E 1220 00:26:45.879 UTC THREAD27: adminhttp
com.twitter.finagle.CancelledRequestException: request cancelled. Remote Info: Not Available
and causing the admin UI to hang, as though it's still waiting to receive a response from Namerd.

I've updated the code handling Consul GET responses in ConsulDtabStore to Try the call to Dtab.read(), and update the activity's state to Activity.Failed when the dtab could not be parsed successfully. Namerd will now log messages which state that the dtab was malformed, and surface the parse error, including showing which character was invalid. The log message now looks like this:

E 1220 20:46:06.770 UTC THREAD31 TraceId:30cbad82deebb433: consul ns consul dtab parsing failed: java.lang.IllegalArgumentException: '/' expected but 'a' found at '[a]sjgfdskjhsdfg;'; dtab: 'asjgfdskjhsdfg;'
The admin UI web interface doesn't currently handle the error in this case, and will load a blank page rather than displaying a useful error message. However, it no longer appears to be awaiting a response, and the error messages in the logs are much more meaningful; and I think changes to the web UI should be in a separate PR.

I've written a new unit test to assert that an unparsable dtab fails the Activity returned by ConsulDtabStore.observe(). This unit test fails against the current master. I've also added some test utility code to make assertions about Activity.State values more ergonomic. I've also manually verified the changes by creating an invalid dtab value in Consul running locally and opening the admin UI for that namespace.