Ignore namerd-provided retry durations
In some situations namerd may, seemingly erroneously, get into a state
where it fails requests with a retry duration that is effectively
infinite, preventing the request from being retried.

In order to track this issue down, this server-provided value is now
ignored in favor of the locally configured backoff.

Additional INFO logging has been added to assist debugging of this
issue.