add access logging for http2 requests (#1786)
We previously did not support access logging for HTTP/2 requests, as the h2 router was experimental until the last release. We've received feature requests on g2 logging from the community.

Add access logging for http2 requests. Logs are stored in ./logs/access.log and can be accessed with: cat ./logs/access.log

Due to the Request and Response object being extensions written by us, the request version is hardcoded as HTTP/2 and we do not have the responseBytes in the log

Fixes: #1776 

Signed-off-by: Franziska von der Goltz <franziska@vdgoltz.eu>