Filter out dirs when listing namespaces in Consul
In order to list namespaces available in Consul KV
ConsulDtabStore makes a call like /kv/some/path/?keys=true that may return not only keys inside /some/path/ but also
an entry for /some/path/ and all dirs inside /some/path/ such as /some/path/foo/ (but not /some/path/foo/bar).
This leads to an exception in API response parsing and fails listing activity so that it's pending forever.

Fixed by omiting the entry for /some/path/ as wells as
all entries ending with / and applying a "validNs" regex.

Added a test imitating a problematic Consul KV structure.

Fixes #1812

Signed-off-by: Borys Pierov <pierovbg@ncbi.nlm.nih.gov>