Backoff before retrying on unexpected erros
Consul namer uses ConsulApi with retry=true to
retry on certain types of errors, and ConsulApi
uses backoffs while retrying. However if error
is not handled by ConsulApi, it is propagated as
exception to SvcAddr and retry is attemted there
without backoff. This is potentially dangerous.

This commit makes SvcAddr to use backoffs stream
provded by ConsulApi.

Fixes #1821

Signed-off-by: Dmytro Kostiuchenko <kostiuchenkod2@ncbi.nlm.nih.gov>