Fix ZooKeeper session management (#1830)
If a ZooKeeper watch is created on a session that expires, a new session will be created but the watch will remain on the expired session. Furthermore, when Namerd encounters ZK errors, it signals that Linkerd should wait for a very long (10 minute) delay before retrying. This leads to extremely slow recovery in the case of an error.

Fix the watch code to follow the new session in when a session expires. Additionally, remove the ability for Namerd to signal retry delays to Linkerd and instead use the backoffs already configured in the client.

Signed-off-by: Alex Leong <alex@buoyant.io>