Do not rely on ConsulApi retries in consul namer (#1827)
Consul namer uses ConsulApi with retry=true to
retry on certain types of errors, backoffs is used
by ConsulApi while retrying. However if error is
not handled by ConsulApi, it is propagated as
exception to SvcAddr and retry is attemted there
without backoff. This is potentially dangerous.

This commit delegates retries fully to SvcAddr and
makes it use backoffs for all retries

Fixes #1821

Signed-off-by: Dmytro Kostiuchenko <edio@archlinux.us>