Update Delegator interface to return a single value only (#1862)
The Delegator interface defines a `delegate` method that returns an Activity.  However, in all cases, only the first value of this Activity was ever used (for the delegator UI).  For complex delegations with fast changing values, a delegation lookup could be quite expensive because many values would be calculated before the Activity was canceled.

We modify the `delegate` method to return a Future instead of an Activity.  This implies changes to all of Namerd's interfaces.

* thriftNamerInterface: We change the client to do a one-shot call with an empty stamp and ignore the stamp in the response instead of long-polling.  For backwards compatibility with long-polling clients, we modify the server to keep requests that have a non-empty stamp pending until cancelled by the client.
* http streaming: No change necessary because the delegate method never took advantage of http streaming.
* mesh: We deprecate the StreamDelegateTree method and modify the implementation to only return a single value on the stream.

I have manually tested all combinations of:
* client version = {1.3.5, HEAD}
* server version = {1.3.5, HEAD}
* interface = {thriftNamerInterface, mesh, http streaming}

Fixes #1846 

Signed-off-by: Alex Leong <alex@buoyant.io>