Fix consul service observation release
Consul namer releases the observation of a service
by raising a specific Exception - ServiceRelease.
Decision whether to release an observation is
taken upon inspection of an observation result
(SvcAddr:67) - if it's a Failure and
ServiceRelease is its imeediate cause, then the
observation should be released. However the cause
may be nested deeply. So the observation is not
always released.

This commit unwraps the root cause of a Thrown
received during observation so we can detect
ServiceRelease even if it was wrapped arbitrary
number of times by underlying code.

Fixed #1820

Signed-off-by: Dmytro Kostiuchenko <edio@archlinux.us>