Fix consul observations leak
LookupCache uses Memoize to return the same instance of Activity
whenever lookup happens in Consul namer. Although returned activity is
backed by an async and lazy Var (backed by polling SvcAddr), it is
itself not async and lazy, as it is a join result of multiple
Acitivities which are not async and lazy. Hence every observer to a
returned activity triggers new SvcAddr creation and, as a result, new
connection to Consul even if Activity is cached.

This commit reworks Activity creation in LookupCache such that it is
indeed async and lazy. Activity remains pending unless all dependencies
to it are satisfied.

Tested in real environment, where previous version of namer was opening
up to 20K connections to consul agent, causing namerd unresponsiveness,
consul crashes. After the fix it's barely 300 connections.

Fixes #1873

Signed-off-by: Dmytro Kostiuchenko <kostiuchenkod2@ncbi.nlm.nih.gov>