Update address set to Neg when consul queries a non-existent datacenter (#1863)
When the ConsulNamer tries to discover a set of addresses using a consul endpoint that contains a non existent dc, it receives a 500 error from consul. The HTTP request is then retried indefinitely even though the request will never result in a valid HTTP response. This causes bind requests to Linkerd to always timeout. The issue is easily observed when using the dtab playground in Linkerd to evaluate paths to address sets and dtabs contain paths that point to non-existent datacenters in consul.

This PR fixes the issue by checking the HTTP 500 message to evaluate if it is a No path to datacenter message. This indicates that there is no dc in consul that contains the provided dc value. In this case, update the service discovery endpoints to Addr.Neg and then retry with a backoff interval.

Change was tested locally through the admin dtab UI and now all dtab requests no longer timeout and give an error back immediately when an invalid dc is used.

fixes #1836

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>