Fix dedup to not leak observations (#1907)
Expiring a client does not necessarily close service discovery watches for that client.  This means, for example, that Linkerd will maintain streaming connections to the Kubernetes API for services that it is no longer routing to and has reaped.  This is due to the use of `Var.apply` in `RichActivity.dedup`.  This variant of `Var.apply` eagerly observes the underlying Event and doesn't release the observation until it is garbage collected.  From the scaladoc for `Var.apply`:

```
   /**
   * Constructs a [[Var]] from an initial value plus an event stream of
   * changes. Note that this eagerly subscribes to the event stream;
   * it is unsubscribed whenever the returned [[Var]] is collected.
   */
```

We change `RichActivity.dedup` to use `Var.async` instead of `Var.apply` in order to do proper reference counting and therefore to only observe the underlying Event while the Activity is observed.

Verified by issuing requests through Linkerd to two different Kubernetes services and then waiting for one of the clients to expire.  Using `netstat`, verify that Linkerd has only 1 connection to the Kubernetes API (previously was 2).

Signed-off-by: Alex Leong <alex@buoyant.io>