cast Http2Stream to Http2FrameStream onStreamClosed (#1908)
This is a follow-up commit for the finagle upgrade. When a stream is closed between Linkerd and an Http/2 server, we see "ClassCastException" in Linkerd. This PR fixes that issue but properly casting a Http2Stream to an Http2FrameStream onStreamClosed

Tests were conducted with current master and this branch. The exception no longer shows up with this fix.

fixes #1905

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>