add a note that highlights transformer and client tls behavior (#1933)
There is a small 'gotcha' users can encounter when using client TLS with kind io.l5d.static with transformers. when linkerd is building client names for routing, transformers present in the linkerd configuration will be prepended to a client's dtab name. At the end of the identification phase, this name will have a different prefix compared to the one set in the client TLS section.

This PR highlights this behavior so that users understand that they may run into this scenario.

fixes #1930

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>