Instrument watches in the io.l5d.namerd interpreter (#1943)
It can be difficult to know which watches the Namerd interpreter has cached, which of them are active, and if they are receiving updates from the Namerd. This has made issues with the Namerd extremely difficult to diagnose and debug.

We instrument the Activities in the Namerd interpreter to keep track of metadata about themselves. This metadata is exposed though an admin endpoint.  We also remove the existing NamerdHandler since this information is now available elsewhere (the dtab interface and this state endpoint).

Fixes #1913 

Signed-off-by: Alex Leong <alex@buoyant.io>