Instrument watches in the io.l5d.mesh interpreter (#1956)
It can be difficult to know which watches the mesh interpreter has cached, which of them are active, and if they are receiving updates from the Namerd. This has made issues with the Namerd extremely difficult to diagnose and debug.

We instrument the Activities in the mesh interpreter to keep track of metadata about themselves. This metadata is exposed though an admin endpoint `/interpreter_state/io.l5d.mesh/default.json`.

Fixes #1914 

Signed-off-by: Alex Leong <alex@buoyant.io>