Configure adding clientId to TTwitter requests
There are Thrift clients for many languages but few support TTwitter
upgrade. Some Thrift servers use the TTwitter clientId header to apply
rate-limiting quotas on clients. Thrift clients that cannot speak
TTwitter are unable to make requests against such servers.

Linkerd already supports upgrading the protocol to the server, even when
the client does not talk TTwitter. This PR adds a Thrift client
configuration option to set the clientId on outgoing TTwitter requests
when no clientId is set on the incoming request. This allows a
TTwitter-unaware client to use a service that requires the clientId be
set by proxying the requests through an appropriately-configured
linkerd.

Signed-off-by: Cory Snider <cory.snider@fitbit.com>