Refactor formatDelegation into DelegateTree.find for easier reuse (#1988)
The `formatDelegation` method returns the delegation steps toward a particular leaf in a delegation tree.  This is generally useful but cannot be reused because it's a private member of `RouterContextFormatter`.

We refactor this method into `DelegateTree` itself, renaming it to `find` so that it can be more easily reused.  We also simplify the method, removing the need for an accumulator and add a test.

Signed-off-by: Alex Leong <alex@buoyant.io>