Handle close terminated HTTP/1.0 responses (#1992)
When Linkerd receives an HTTP/1.0 response that is terminated by closing the connection, the response will not have a Content-Length header.  This causes Linkerd to interpret the response as chunk encoded.  Since Linkerd keeps persistent connections open to its clients, clients will not have a content-length header or connection termination to signal the end of the response.  This causes them to hang, waiting for the response to complete.

This issue was introduced in #1904 when we stopped upgrading responses to HTTP/1.1.

For HTTP/1.0 responses with no transfer-coding and no content-length, we fully buffer the response and set the content-length header. 

Fixes #1978 

Signed-off-by: Alex Leong <alex@buoyant.io>