Disable statistics exported by the JVM for hsprefdata
The JVM by default exports statistics by mmap-ing a file in /tmp (hsperfdata).
On Linux, modifying a memory mapped file can block until disk I/O completes, which can be hundreds of milliseconds.
Since the JVM modifies these statistics during garbage collection and safepoints, this causes pauses that are hundreds of milliseconds long on spinning disks.

This will break tools that read this file, such as jcmd, jstat, jconsole. If you *really* need those, and you are running /tmp backed by
tmpfs or an SSD, it's okay to enable this feature.

References:

http://www.evanjones.ca/jvm-mmap-pause.html
https://engineering.linkedin.com/blog/2016/02/eliminating-large-jvm-gc-pauses-caused-by-background-io-traffic

Note: this is a problem for writing gc logs as well (out of scope), so make sure if you are enabling GC logs for Linkerd, to write them to an SSD or tmpfs.

Fixes #1996

Signed-off-by: Chris Goffinet <chris@threecomma.io>