Rework H2 stream cancellation (#1934)
There does not exist a reliable mechanism for the consumer of an H2 stream to cancel/reset it.  For example, consider a client has received a response stream from the server but wishes to terminate the stream without reading to the end.  The client can attempt to read a frame from the response stream and immediately interrupt the read, which may reset the stream, but this is timing dependent and will not always succeed.  The client may also use the request stream to issue a reset but this is awkward because the caller may not have kept a reference to the request stream, especially if the request stream is complete.  

We refactor the Stream interface to remove `reset` from `Stream.Writer` and add `cancel` to `Stream`, as well as an `onCancel` Future.  Consumers of a Stream may call `cancel` to indicate that they wish to terminate the stream.  Producers to the stream may listen to the `onCancel` Future to be notified of a consumer initiated stream termination.

We also substantially reworked Netty4StreamTransport to make the handling of resets clearer.  Rather than having dependently nested stream states, we now keep an independent state for the send (local) side and the recv (remote) side.  We also centralized the reset logic as much as possible into a `reset` method.  

This change unlocks further improvements to the interface for cancellation in gRPC.  In turn that will allow us to add proper cancellation to the io.l5d.mesh gRPC API to address #1906 

Signed-off-by: Alex Leong <alex@buoyant.io>