Be part of an opentracing context originated externally
We wanted linkerd to continue a tracing context created outside (in our load balancers basically) instead of creating a new context and using their own headers when propagating it.

To solve it, we have added a condition in the method that parses headers to look for a span header, and if present, parse that information and use it to create the tracing context. Also, we have added the headers to propagate outside linkerd, to make life easier for applications that want to use opentracing libraries.

We have validated this in our environment, using linkerd as service mesh as part of a DC/OS installation. We have seen the same traceId in zipkin with this change for both LB and linkerd.

Fixes #[1447]