Change ordering of the server stack (#1998)
The linkerd provides an option to turn off the error messages via the `contextClear` setting. Even with `contextClear` setting set we get verbose errors from linkerd. This happens because of the ordering of the server stack. Exceptions are turned into error responses in the `ErrorResponder.module` and bodies are scrubbed by the `Headers.Ctx.serverModule`. But because bodies are scrubbed before exceptions are turned into error responses, the error bodies can escape. `Headers.Ctx.serverModule` needs to be before the `http.ErrorResponder.module` so that errors responses from the ErrorResponder will be cleared when `clearContext` is set.

Fixes: #1163 

Signed-off-by: Mantas Stoškus <mantas@inventi.lt>